+++ test/unit/visualization/test_circuit_schedule.py
@@ -16,8 +16,8 @@ import ddt
 import numpy as np
 from qiskit_ibm_runtime.visualization.utils import plotly_module
 from qiskit_ibm_runtime.visualization.circuit_schedule import CircuitSchedule
-from ..ibm_test_case import IBMTestCase
-from .mock.fake_circuit_schedule_timing import FakeCircuitScheduleInputData
+from ...ibm_test_case import IBMTestCase
+from ..mock.fake_circuit_schedule_timing import FakeCircuitScheduleInputData
 
 
 @ddt.ddt
